+++ packages/ts-newskit/src/components/navigation/global-nav/topnav/createMenus.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { Menu, MenuItem, MenuSub, MenuDivider, Visible } from 'newskit';
 import { AccountMenu } from '../styles';
+import { MenuItemParent } from '../types';
 
 const menuItemPresets = {
   minHeight: '60px',
@@ -9,8 +10,16 @@ const menuItemPresets = {
   typographyPreset: 'topNav010'
 };
 
-export const createMenu = (menuData: any) => {
+export const createMenu = (
+  menuData: MenuItemParent[],
+  clickHandler: (title: string) => void
+) => {
   const [moreSelected, setMoreSelected] = useState<boolean>(false);
+
+  const handleMoreClick = () => {
+    setMoreSelected(!moreSelected);
+    clickHandler('More');
+  };
 
   const moreMenuLength = menuData.length - 4;
   const navItems = menuData
@@ -28,6 +37,7 @@ export const createMenu = (menuData: any) => {
             md: 'menuItem'
           }
         }}
+        onClick={() => clickHandler(title)}
         key={url}
       >
         {title}
@@ -40,7 +50,7 @@ export const createMenu = (menuData: any) => {
       <Visible lg>
         <MenuSub
           title="More"
-          onClick={() => setMoreSelected(!moreSelected)}
+          onClick={handleMoreClick}
           selected={moreSelected}
           expanded={moreSelected}
           overrides={{
@@ -54,7 +64,7 @@ export const createMenu = (menuData: any) => {
             overrides={{ spaceInline: 'sizing000' }}
             aria-label="More menu items"
           >
-            {createMoreMenu(menuData, moreMenuLength)}
+            {createMoreMenu(menuData, moreMenuLength, clickHandler)}
           </Menu>
         </MenuSub>
       </Visible>
@@ -62,7 +72,11 @@ export const createMenu = (menuData: any) => {
   );
 };
 
-const createMoreMenu = (menuData: any, moreMenuLength: number) =>
+const createMoreMenu = (
+  menuData: MenuItemParent[],
+  moreMenuLength: number,
+  clickHandler: (title: string) => void
+) =>
   menuData
     .slice(-moreMenuLength)
     .map(({ title, url }: { title: string; url: string }) => (
@@ -74,12 +88,17 @@ const createMoreMenu = (menuData: any, moreMenuLength: number) =>
           typographyPreset: 'topNav010'
         }}
         key={url}
+        onClick={() => clickHandler(`more: ${title}`)}
       >
         {title}
       </MenuItem>
     ));
 
-export const createAccountMenu = (isLoggedIn: boolean, menuData: any) => {
+export const createAccountMenu = (
+  isLoggedIn: boolean,
+  menuData: MenuItemParent[],
+  clickHandler: (title: string) => void
+) => {
   const [myAccountSelected, setMyAccountSelected] = useState<boolean>(false);
 
   return isLoggedIn ? (
@@ -90,13 +109,17 @@ export const createAccountMenu = (isLoggedIn: boolean, menuData: any) => {
         overrides={{
           ...menuItemPresets
         }}
+        onClick={() => clickHandler('Times +')}
       >
         Times +
       </MenuItem>
       <MenuDivider />
       <MenuSub
         title="My Account"
-        onClick={() => setMyAccountSelected(!myAccountSelected)}
+        onClick={() => {
+          setMyAccountSelected(!myAccountSelected);
+          clickHandler('My Account');
+        }}
         selected={myAccountSelected}
         expanded={myAccountSelected}
         overrides={{
@@ -120,6 +143,7 @@ export const createAccountMenu = (isLoggedIn: boolean, menuData: any) => {
                 typographyPreset: 'topNav010'
               }}
               key={url}
+              onClick={() => clickHandler(title)}
             >
               {title}
             </MenuItem>
@@ -133,7 +157,11 @@ export const createAccountMenu = (isLoggedIn: boolean, menuData: any) => {
       aria-label="My Account Menu"
     >
       <MenuDivider />
-      <MenuItem href="/login" overrides={{ ...menuItemPresets }}>
+      <MenuItem
+        href="/login"
+        overrides={{ ...menuItemPresets }}
+        onClick={() => clickHandler('Login')}
+      >
         Login
       </MenuItem>
       <MenuDivider />
@@ -144,6 +172,7 @@ export const createAccountMenu = (isLoggedIn: boolean, menuData: any) => {
           typographyPreset: 'topNav010',
           spaceInline: 'space090'
         }}
+        onClick={() => clickHandler('Subscribe')}
       >
         Subscribe
       </MenuItem>