+++ packages/article-magazine-comment/CHANGELOG.md
@@ -3,6 +3,110 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.74.31](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.30...@times-components/article-magazine-comment@3.74.31) (2023-12-14)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.30](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.29...@times-components/article-magazine-comment@3.74.30) (2023-12-12)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.29](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.28...@times-components/article-magazine-comment@3.74.29) (2023-12-08)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.28](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.27...@times-components/article-magazine-comment@3.74.28) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.27](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.26...@times-components/article-magazine-comment@3.74.27) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.26](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.25...@times-components/article-magazine-comment@3.74.26) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.25](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.24...@times-components/article-magazine-comment@3.74.25) (2023-12-06)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.24](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.23...@times-components/article-magazine-comment@3.74.24) (2023-12-05)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.23](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.22...@times-components/article-magazine-comment@3.74.23) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.22](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.21...@times-components/article-magazine-comment@3.74.22) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.21](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.20...@times-components/article-magazine-comment@3.74.21) (2023-12-01)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.20](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.19...@times-components/article-magazine-comment@3.74.20) (2023-11-30)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.74.19](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.18...@times-components/article-magazine-comment@3.74.19) (2023-11-29)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
 ## [3.74.18](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.74.17...@times-components/article-magazine-comment@3.74.18) (2023-11-29)
 
 **Note:** Version bump only for package @times-components/article-magazine-comment