+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,110 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.73.31](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.30...@times-components/article-main-comment@2.73.31) (2023-12-14)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.30](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.29...@times-components/article-main-comment@2.73.30) (2023-12-12)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.29](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.28...@times-components/article-main-comment@2.73.29) (2023-12-08)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.28](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.27...@times-components/article-main-comment@2.73.28) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.27](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.26...@times-components/article-main-comment@2.73.27) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.26](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.25...@times-components/article-main-comment@2.73.26) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.25](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.24...@times-components/article-main-comment@2.73.25) (2023-12-06)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.24](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.23...@times-components/article-main-comment@2.73.24) (2023-12-05)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.23](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.22...@times-components/article-main-comment@2.73.23) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.22](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.21...@times-components/article-main-comment@2.73.22) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.21](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.20...@times-components/article-main-comment@2.73.21) (2023-12-01)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.20](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.19...@times-components/article-main-comment@2.73.20) (2023-11-30)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
+## [2.73.19](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.18...@times-components/article-main-comment@2.73.19) (2023-11-29)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.73.18](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.73.17...@times-components/article-main-comment@2.73.18) (2023-11-29)
 
 **Note:** Version bump only for package @times-components/article-main-comment