+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,113 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.116.15](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.14...@times-components/article-skeleton@1.116.15) (2023-12-14)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.14](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.13...@times-components/article-skeleton@1.116.14) (2023-12-12)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.13](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.12...@times-components/article-skeleton@1.116.13) (2023-12-08)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.12](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.11...@times-components/article-skeleton@1.116.12) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.11](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.10...@times-components/article-skeleton@1.116.11) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.10](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.9...@times-components/article-skeleton@1.116.10) (2023-12-07)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.9](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.8...@times-components/article-skeleton@1.116.9) (2023-12-06)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.8](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.7...@times-components/article-skeleton@1.116.8) (2023-12-05)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.7](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.6...@times-components/article-skeleton@1.116.7) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.6](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.5...@times-components/article-skeleton@1.116.6) (2023-12-04)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.5](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.4...@times-components/article-skeleton@1.116.5) (2023-12-01)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+## [1.116.4](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.3...@times-components/article-skeleton@1.116.4) (2023-11-30)
+
+
+### Bug Fixes
+
+* **TMRX-1550:** newsletter puff not appearing in irish sport section ([#3676](https://github.com/newsuk/times-components/issues/3676)) ([b72ab65](https://github.com/newsuk/times-components/commit/b72ab65bee181b305b7f382bd6f50c527e066ff7))
+
+
+
+
+
+## [1.116.3](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.2...@times-components/article-skeleton@1.116.3) (2023-11-29)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
 ## [1.116.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.116.1...@times-components/article-skeleton@1.116.2) (2023-11-29)
 
 **Note:** Version bump only for package @times-components/article-skeleton